bug fixes and updates

+++ src/rudolph/geometry.cpp
@@ -18,7 +18,7 @@ void Point2D::translate(int dx, int dy) {
 }
 
 void Point2D::scale(int sx, int sy) {
-    Matrix<int> t(3, 3);
+    Matrix<double> t(3, 3);
     t(0, 0) = sx;
     t(1, 1) = sy;
     t(2, 2) = 1;
@@ -29,7 +29,7 @@ void Point2D::scale(int sx, int sy) {
 void Point2D::rotate(double angle) {
     const double pi = std::acos(-1);
 
-    Matrix<int> t(3, 3);
+    Matrix<double> t(3, 3);
     t(0, 0) = std::cos(angle*pi/180);
     t(0, 1) = -std::sin(angle*pi/180);
     t(1, 0) = std::sin(angle*pi/180);
